pyproject.toml
[tool.poetry]
name = "hh-vacancies-project"
version = "0.1.0"
description = ""
authors = ["gvia <gritcaev.vv@gmail.com>"]
readme = "README.md"
packages = [{include = "hh_vacancies_project"}]

[tool.poetry.dependencies]
python = "^3.13"


[tool.poetry.group.lint.dependencies]
flake8 = "^7.1.1"
mypy = "^1.14.1"
black = "^25.1.0"
isort = "^6.0.1"

[tool.poetry.group.dev.dependencies]
requests = "^2.32.3"
dotenv = "^0.9.9"
psycopg2-binary = "^2.9.10"

[tool.mypy]
disallow_untyped_defs = true
no_implicit_optional = true
warn_return_any = true
exclude = "venv"

[tool.black]
line-length = 119
exclude = ".git"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
